UBSan Extension
In this revision, we enable generating code for  runtime checks which make use of the must-not-alias predicates generated, to identify the presence of  Undefined Behaviour due to OOE non-determinism, in the code. LLVM’s widening of bitfield objects (objects which are smaller than a single byte) to their byte-level objects, may result in the generation of unsound predicates during the lowering to LLVM IR because two different bitfields that belong to the same byte would now have identical addresses and our predicate would incorrectly encode a must-not-alias relationship between the identical byte addresses. Thus, to preserve soundness, we conservatively ignore all predicates both of whose expressions are bitfield types. We also ignore those predicates which have at least one function tag associated to them as for these predicates, the must-not-alias assertions is conditioned upon the functions being `readnone` and inferring that during code generation is not always possible, due to the unavailability of the LLVM function summaries at that time.