Unseq Alias Visitor
In this revision, we implement a plugin called UnsequencedAliasVisitor, which runs our algorithm on the AST before the lowering to LLVM IR happens. As per our algorithm, we also need the information about function summaries in order to know whether a function is readnone. However, we do not have function summaries available at this point. To implement our algorithm, we tag each predicate with the function calls that appear in either expression of the predicate. This information will be subsequently used while LLVM IR generation and alias analysis.

We use a Predicate Map that maps each top-level expression in the AST to a list of corresponding tagged must-not-alias predicates. Each must not alias predicate is represented as a tuple containing a serial no., the pair of pointers which must not alias and the list of function calls as described above.